# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer API Documentation",
        "description": "Documentation de l'API Customer",
        "contact": {
            "email": "support@example.com"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/customer-canvassing-steps": {
            "get": {
                "tags": [
                    "CustomerCanvassingSteps"
                ],
                "summary": "Liste toutes les étapes de démarchage clients",
                "operationId": "e61cdbc33f4d905b0de13879337d32e6",
                "responses": {
                    "200": {
                        "description": "Liste des étapes de démarchage clients récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CustomerCanvassingStep"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des étapes de démarchage clients récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CustomerCanvassingSteps"
                ],
                "summary": "Crée une nouvelle étape de démarchage client",
                "operationId": "9607ca5273add357142976bdf9dcbdc9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCustomerCanvassingStepRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Étape de démarchage client créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerCanvassingStep"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Étape de démarchage client créée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/customer-canvassing-steps/{id}": {
            "get": {
                "tags": [
                    "CustomerCanvassingSteps"
                ],
                "summary": "Affiche les détails d'une étape de démarchage client spécifique",
                "operationId": "58485a09f2773d2145c4ced3a5ac4e53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étape de démarchage client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Étape de démarchage client récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerCanvassingStep"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Étape de démarchage client récupérée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Étape de démarchage client non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "CustomerCanvassingSteps"
                ],
                "summary": "Met à jour une étape de démarchage client existante",
                "operationId": "ebb94efbca14bc704c1e48dc7bdf0fbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étape de démarchage client à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomerCanvassingStepRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Étape de démarchage client mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerCanvassingStep"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Étape de démarchage client mise à jour avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Étape de démarchage client non trouvée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "CustomerCanvassingSteps"
                ],
                "summary": "Supprime une étape de démarchage client",
                "operationId": "8a0ba590ad35ff1de42f9e815ee2389c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étape de démarchage client à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Étape de démarchage client supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Étape de démarchage client supprimée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Étape de démarchage client non trouvée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCustomerCanvassingStepRequest": {
                "required": [
                    "name",
                    "order"
                ],
                "properties": {
                    "name": {
                        "description": "Nom de l'étape",
                        "type": "string",
                        "maxLength": 200,
                        "example": "Premier contact"
                    },
                    "order": {
                        "description": "Ordre d'affichage",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateCustomerCanvassingStepRequest": {
                "properties": {
                    "name": {
                        "description": "Nom de l'étape",
                        "type": "string",
                        "maxLength": 200,
                        "example": "Premier contact"
                    },
                    "order": {
                        "description": "Ordre d'affichage",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "CustomerCanvassingStep": {
                "title": "CustomerCanvassingStep",
                "description": "Modèle d'étape de démarchage client",
                "properties": {
                    "id_customer_canvassing_step": {
                        "description": "ID unique de l'étape de démarchage",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom de l'étape",
                        "type": "string",
                        "maxLength": 200,
                        "example": "Premier contact"
                    },
                    "order": {
                        "description": "Ordre d'affichage",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière modification",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}